+++ functionRename/Services/LikeCommentService.csx
@@ -8,6 +8,9 @@ public class LikeCommentService
     public string Like(string sessionId)
     {
         var revisionId = new RevisionService().CreateIfDoesntExist(sessionId);
+        if (revisionId == Guid.Empty)
+            return "You need to authorise the user";
+
         var newLiker = new Authorisation().GetUser(new Registration().Generate(), sessionId, Actions.LikeYourRevision);
         var userQuery = new UserModelQuery();
         userQuery.Add(newLiker);
@@ -24,6 +27,9 @@ public class LikeCommentService
     public string Dislike(string sessionId)
     {
         var myLikers = new UserModelQuery().GetList(sessionId, Actions.LikeYourRevision);
+        if (myLikers.Count == 0)
+            return "You have no likers";
+
         var dislike = new Revisions().Dislike(myLikers);
         new UserModelQuery().Delete(myLikers);
         return "Disliked";
@@ -32,23 +38,31 @@ public class LikeCommentService
     public string LeftComment(string sessionId, bool withMentionMe = false)
     {
         var me = new UserModelQuery().Get(sessionId, Actions.Me);
-        var revisionId = new RevisionService().CreateIfDoesntExist(sessionId);
-        var newCommenter = new Authorisation().GetUser(new Registration().Generate(), sessionId, Actions.CommentYourRevision);
-        var userQuery = new UserModelQuery();
-        userQuery.Add(newCommenter);
-        var comment = withMentionMe ? new Revisions().LeftComment(revisionId, newCommenter.AccessToken, '@' + me.UserName) : new Revisions().LeftComment(revisionId, newCommenter.AccessToken);
-        if (comment != null)
-        {
-            userQuery.Update(newCommenter.PartitionKey, newCommenter.RowKey, commentId:comment);
-            return "Commented";
+        if(me != null)
+        { 
+            var revisionId = new RevisionService().CreateIfDoesntExist(sessionId);
+            var newCommenter = new Authorisation().GetUser(new Registration().Generate(), sessionId, Actions.CommentYourRevision);
+            var userQuery = new UserModelQuery();
+            userQuery.Add(newCommenter);
+            var comment = withMentionMe ? new Revisions().LeftComment(revisionId, newCommenter.AccessToken, '@' + me.UserName) : new Revisions().LeftComment(revisionId, newCommenter.AccessToken);
+            if (comment != null)
+            {
+                userQuery.Update(newCommenter.PartitionKey, newCommenter.RowKey, commentId:comment);
+                return "Commented";
+            }
+            else
+                return "Something went wrong";
         }
         else
-            return "Something went wrong";
+            return "You need to authorise the user";
     }
 
     public string DeleteComment(string sessionId)
     {
         var myCommenters = new UserModelQuery().GetList(sessionId, Actions.CommentYourRevision);
+        if (myCommenters.Count == 0)
+            return "Nobody comments your revision";
+
         var deleteComment = new Revisions().DeleteComment(myCommenters);
         new UserModelQuery().Delete(myCommenters);
         return "Comment deleted";
